pyproject.toml
[project]
name = "local-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "auto-gptq>=0.7.1",
    "flask>=3.1.2",
    "langchain>=1.1.2",
    "langchain-core>=1.1.1",
    "langchain-text-splitters>=1.0.0",
    "loguru>=0.7.3",
    "optimum>=2.0.0",
    "psycopg2-binary>=2.9.11",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
